Upgraded packages and add Vite - Updated old packages - Replaced Babel and Rollup with Vite - Replaced Ava with Vitest - Renamed testing files to use Vitest defaults
