--- babel.config.js
@@ -1,18 +1,0 @@
-module.exports = function (api) {
-  api.cache(true);
-  return {
-    presets: [
-      [
-        "@babel/preset-env",
-        {
-          useBuiltIns: "usage",
-          corejs: 3,
-        },
-      ],
-      "@babel/preset-typescript",
-    ],
-    plugins: ["@babel/plugin-transform-runtime"],
-    comments: false,
-    exclude: [/(\/|\\)core-js(\/|\\)/],
-  };
-};
