--- src/index.js
@@ -1,190 +1,0 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g = Object.create((typeof Iterator === "function" ? Iterator : Object).prototype);
-    return g.next = verb(0), g["throw"] = verb(1), g["return"] = verb(2), typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (g && (g = 0, op[0] && (_ = 0)), _) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-require("@k2oss/k2-broker-core");
-metadata = {
-    systemName: "com.k2.example",
-    displayName: "Example Broker",
-    description: "An example broker that accesses JSONPlaceholder.",
-};
-ondescribe = function (_a) {
-    return __awaiter(this, arguments, void 0, function (_b) {
-        var configuration = _b.configuration;
-        return __generator(this, function (_c) {
-            postSchema({
-                objects: {
-                    todo: {
-                        displayName: "TODO",
-                        description: "Manages a TODO list",
-                        properties: {
-                            id: {
-                                displayName: "ID",
-                                type: "number",
-                            },
-                            userId: {
-                                displayName: "User ID",
-                                type: "number",
-                            },
-                            title: {
-                                displayName: "Title",
-                                type: "string",
-                            },
-                            completed: {
-                                displayName: "Completed",
-                                type: "boolean",
-                            },
-                        },
-                        methods: {
-                            get: {
-                                displayName: "Get TODO",
-                                type: "read",
-                                inputs: ["id"],
-                                outputs: ["id", "userId", "title", "completed"],
-                            },
-                            getParams: {
-                                displayName: "Get TODO",
-                                type: "read",
-                                parameters: {
-                                    pid: {
-                                        displayName: "param1",
-                                        description: "Description Of Param 1",
-                                        type: "number",
-                                    },
-                                },
-                                requiredParameters: ["pid"],
-                                outputs: ["id"],
-                            },
-                        },
-                    },
-                },
-            });
-            return [2 /*return*/];
-        });
-    });
-};
-onexecute = function (_a) {
-    return __awaiter(this, arguments, void 0, function (_b) {
-        var _c;
-        var objectName = _b.objectName, methodName = _b.methodName, parameters = _b.parameters, properties = _b.properties, configuration = _b.configuration, schema = _b.schema;
-        return __generator(this, function (_d) {
-            switch (_d.label) {
-                case 0:
-                    _c = objectName;
-                    switch (_c) {
-                        case "todo": return [3 /*break*/, 1];
-                    }
-                    return [3 /*break*/, 3];
-                case 1: return [4 /*yield*/, onexecuteTodo(methodName, properties, parameters)];
-                case 2:
-                    _d.sent();
-                    return [3 /*break*/, 4];
-                case 3: throw new Error("The object " + objectName + " is not supported.");
-                case 4: return [2 /*return*/];
-            }
-        });
-    });
-};
-function onexecuteTodo(methodName, properties, parameters) {
-    return __awaiter(this, void 0, void 0, function () {
-        var _a;
-        return __generator(this, function (_b) {
-            switch (_b.label) {
-                case 0:
-                    _a = methodName;
-                    switch (_a) {
-                        case "get": return [3 /*break*/, 1];
-                        case "getParams": return [3 /*break*/, 3];
-                    }
-                    return [3 /*break*/, 5];
-                case 1: return [4 /*yield*/, onexecuteTodoGet(properties)];
-                case 2:
-                    _b.sent();
-                    return [3 /*break*/, 6];
-                case 3: return [4 /*yield*/, onexecuteTodoGetWithParams(parameters)];
-                case 4:
-                    _b.sent();
-                    return [3 /*break*/, 6];
-                case 5: throw new Error("The method " + methodName + " is not supported.");
-                case 6: return [2 /*return*/];
-            }
-        });
-    });
-}
-function onexecuteTodoGet(properties) {
-    return new Promise(function (resolve, reject) {
-        var xhr = new XMLHttpRequest();
-        xhr.onreadystatechange = function () {
-            try {
-                if (xhr.readyState !== 4)
-                    return;
-                if (xhr.status !== 200)
-                    throw new Error("Failed with status " + xhr.status);
-                var obj = JSON.parse(xhr.responseText);
-                postResult({
-                    id: obj.id,
-                    userId: obj.userId,
-                    title: obj.title,
-                    completed: obj.completed,
-                });
-                resolve();
-            }
-            catch (e) {
-                reject(e);
-            }
-        };
-        if (typeof properties["id"] !== "number")
-            throw new Error('properties["id"] is not of type number');
-        xhr.open("GET", "https://jsonplaceholder.typicode.com/todos/" +
-            encodeURIComponent(properties["id"]));
-        xhr.setRequestHeader("test", "test value");
-        xhr.send();
-    });
-}
-function onexecuteTodoGetWithParams(parameters) {
-    return new Promise(function (resolve, reject) {
-        try {
-            postResult({
-                id: parameters["pid"],
-            });
-            resolve();
-        }
-        catch (e) {
-            reject(e);
-        }
-    });
-}
