+++ tsconfig.json
@@ -1,10 +1,16 @@
 {
   "extends": "@k2oss/k2-broker-core/shared/tsconfig.json",
   "compilerOptions": {
-    "module": "CommonJS",
-    "lib": [
-      "es6",
-      "es2015.promise"
-    ]
-  }
+    "target": "ESNext",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "outDir": "dist",
+    "declaration": true,
+    "declarationDir": "dist/types",
+    "strict": false,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+  },
+  "include": ["src"],
+  "exclude": ["node_modules"]
 }
