+++ vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+    test: {
+        coverage: {
+            provider: 'istanbul' // or 'v8'
+        },
+    },
+})
